+++ packages/vda-node-manager/CHANGELOG.md
@@ -1,3 +1,9 @@
+2024-06-28 (v)
+-------------------
+
+- Add a dependency to the `@verida/vda-client-base`
+- Update test codes for refactoring DIDs
+
 2024-03-22 (v3.0.2)
 -------------------